Started building a tool to build test fixtures.
Prior to this, all files in Dropbox had to be manually prepared in order
to record a VCR cassette and build the right test.

This commit adds a DropboxScaffoldBuilder which can be used to generate
the needed files in Dropbox for each endpoint. This can be invoked from
a rake task.

This is related to #15.